migrate to new sdr-server api

+++ src/sdr/sdr_server_api.h
@@ -36,7 +36,7 @@ struct sdr_server_request {
 
 struct sdr_server_response {
 	uint8_t status;
-	uint8_t details; // on success contains file index, on error contains error code
+    uint32_t details; // on success contains file index, on error contains error code
 } __attribute__((packed));
